[Analysis] Extract KnownFPClass
