+++ llvm/utils/gn/secondary/llvm/lib/Support/BUILD.gn
@@ -102,6 +102,7 @@ static_library("Support") {
     "IntervalMap.cpp",
     "JSON.cpp",
     "KnownBits.cpp",
+    "KnownFPClass.cpp",
     "LEB128.cpp",
     "LineIterator.cpp",
     "Locale.cpp",
